refactor: Extract PageTableEntry and XWRMode into page_table_entry.rs (R-02)
Split the 833-line page_tables.rs monolith by moving PageTableEntry
struct, XWRMode enum, and their unit tests into a dedicated module.
PageTable and RootPageTableHolder remain in page_tables.rs.
